Load data from URL instead of file
Currently the `load_data.py` script loads data from a file, but we'd
like to load it from a URL instead.

To do this, I added subclassed the LoadData class and overrode the
get_raw_data() method to retrieve the data from a URL instead of the
file.

I used urllib.request.urlopen() to fetch the data as it's part of the
standard library and it didn't seem worth it to install a package such
as requests just to use in one place when the standard library will do
fine.