Store capital and top level domains for countries
In the data provided, there's just one capital per country, so an extra
field was added to the country table.

For top level domain names (TLDs) there were countries with multiple
TLDs, so this was added as an additional table.

I wasn't quite sure if the load_data.py script should also update the
existing data, but I made it possible to do so with the `--update`
argument.

I didn't worry about transactions for the moment, but that might be a
concern when updating the TLDs for each country if there is concurrent
access to the database.